rm global import of wirefree
