+++ mio/devices/__init__.py
@@ -4,5 +4,6 @@ Devices!
 
 from miniscope_io.devices.camera import Camera, Miniscope
 from miniscope_io.devices.device import Device
+from miniscope_io.devices.wirefree import WireFreeMiniscope
 
-__all__ = ["Camera", "Device", "Miniscope"]
+__all__ = ["Camera", "Device", "Miniscope", "WireFreeMiniscope"]
