+++ mio/exceptions.py
@@ -9,14 +9,14 @@ class ConfigurationError(ValueError):
 
 class InvalidSDException(Exception):
     """
-    Raised when :class:`.io.SDCard` is used with a drive that doesn't have the
+    Raised when :class:`.io.WireFreeMiniscope` is used with a drive that doesn't have the
     appropriate WRITE KEYS in its header
     """
 
 
 class EndOfRecordingException(StopIteration):
     """
-    Raised when :class:`.io.SDCard` is at the end of the available recording!
+    Raised when :class:`.io.WireFreeMiniscope` is at the end of the available recording!
     """
