+++ tests/test_io.py
@@ -7,8 +7,8 @@ import csv
 import numpy as np
 import warnings
 
+from mio import WireFreeMiniscope
 from mio.models.sdcard import SDBufferHeader
-from mio.io import SDCard
 from mio.io import BufferedCSVWriter
 from mio.exceptions import EndOfRecordingException
 from mio.models.data import Frame
@@ -173,7 +173,7 @@ def test_relative_path():
     rel_path = abs_child.relative_to(abs_cwd)
 
     assert not rel_path.is_absolute()
-    sdcard = SDCard(drive=rel_path, layout="wirefree-sd-layout")
+    sdcard = WireFreeMiniscope(drive=rel_path, layout="wirefree-sd-layout")
 
     # check we can do something basic like read config
     assert sdcard.config is not None
@@ -184,7 +184,7 @@ def test_relative_path():
     # now try with an absolute path
     abs_path = rel_path.resolve()
     assert abs_path.is_absolute()
-    sdcard_abs = SDCard(drive=abs_path, layout="wirefree-sd-layout")
+    sdcard_abs = WireFreeMiniscope(drive=abs_path, layout="wirefree-sd-layout")
     assert sdcard_abs.config is not None
     assert sdcard_abs.drive.is_absolute()
 
@@ -216,7 +216,7 @@ def test_to_img(wirefree_battery, n_frames, hash, tmp_path):
 
     assert out_hash == hash
 
-    sd = SDCard(out_file, "wirefree-sd-layout-battery")
+    sd = WireFreeMiniscope(out_file, "wirefree-sd-layout-battery")
 
     # we should be able to read all the frames!
     frames = []
